refactor: correct namespace structure
