+++ src/MiniCSharpCompiler.Core/Lexer/Lexer.cs
@@ -1,13 +1,12 @@
 using Microsoft.CodeAnalysis;
 using MiniCSharpCompiler.Core.Interfaces;
 
-namespace MiniCSharpCompiler.Core.Lexer
+namespace MiniCSharpCompiler.Core.Lexer;
+
+public class Lexer : ILexer
 {
-    public class Lexer : ILexer
+    public IEnumerable<SyntaxToken> Tokenize(string sourceCode)
     {
-        public IEnumerable<SyntaxToken> Tokenize(string sourceCode)
-        {
-            throw new NotImplementedException();
-        }
+        throw new NotImplementedException();
     }
 }
